Move components shared across pages into share folder. Add temp mocks for image imports for unit testing
