+++ src/ui/components/shared/DeviceCard.js
