+++ src/ui/components/shared/DeviceGrid.js
